refactor: Complete Phase 2 high-priority component modularization
- Split admin page (687→457 lines, 33% reduction) into AdminDashboard, AdminTabNavigation, and AdminTabContent components
- Refactor ProductAddition (434→85 lines, 80% reduction) into CategoryAdditionForm, ProductAdditionForm, and ProductAdditionInfo
- Modularize category page (417→79 lines, 81% reduction) with useCategoryData hook, CategoryHero, CategoryProductGrid, and state components
- Extract reusable loading and error state components for consistent UX
- Improve TypeScript type safety across all refactored components
- Maintain full functionality while significantly improving maintainability
