+++ src/app/admin/page.tsx
@@ -5,24 +5,10 @@ import { useAdminBackups } from '@/lib/hooks/useAdminBackups';
 import { useAdminCoupons } from '@/lib/hooks/useAdminCoupons';
 import { useAdminProducts } from '@/lib/hooks/useAdminProducts';
 import { useAdminSession } from '@/lib/hooks/useAdminSession';
-import {
-    GlassButton,
-    GlassCard,
-    GlassIcon,
-    GlassInput,
-    GlassContainer,
-    NeonBubblesBackground
-} from "@/app/components/Glassmorphism";
-import Navbar from "@/app/components/Navbar/Navbar";
-import Heading from "@/app/components/common/Heading";
-import StockStatusBadge from "@/app/components/common/StockStatusBadge";
+import { GlassCard, GlassIcon, GlassInput } from "@/app/components/Glassmorphism";
 import AdminLayout from "@/app/components/admin/AdminLayout";
-import BackupManagement from "@/app/components/admin/BackupManagement";
-import PriceManagement from "@/app/components/admin/PriceManagement";
-import InventoryManagement from "@/app/components/admin/InventoryManagement";
-import ProductAddition from "@/app/components/admin/ProductAddition";
-import CouponManagement from "@/app/components/admin/CouponManagement";
-
+import AdminTabNavigation from "@/app/components/admin/AdminTabNavigation";
+import AdminTabContent from "@/app/components/admin/AdminTabContent";
 
 // Session timeout (30 minutes)
 const SESSION_TIMEOUT = 30 * 60 * 1000;
@@ -48,85 +34,71 @@ export default function AdminPage() {
     setPassword,
     loginAttempts,
     isLocked,
+    lockTimeRemaining,
     message,
     setMessage,
     isLoading,
     setIsLoading,
     handleLogin,
-    handleLogout,
-    startSessionTimer,
-    clearSessionTimer,
-    setIsAuthenticated
+    handleLogout
   } = session;
 
   // Use the custom products hook
   const productsApi = useAdminProducts({ isAuthenticated, setMessage, username, password });
   const {
     products,
-    setProducts,
     categories,
-    setCategories,
     selectedCategory,
     setSelectedCategory,
-    categoryProducts,
-    setCategoryProducts,
     selectedProductId,
     setSelectedProductId,
-    productForm,
-    setProductForm,
     categoryForm,
     setCategoryForm,
+    productForm,
+    setProductForm,
     newDimensionField,
     setNewDimensionField,
     newFeature,
     setNewFeature,
     isAddLoading,
-    setIsAddLoading,
     isUploading,
-    setIsUploading,
     selectedImage,
-    setSelectedImage,
     imagePreview,
-    setImagePreview,
-    loadProducts,
     addCategory,
     addProduct,
     addDimensionField,
     removeDimensionField,
+    handleImageSelect,
     handleImageUpload,
-    handleImageSelect
+    loadProducts
   } = productsApi;
 
   // Use the custom backups hook
   const backupsApi = useAdminBackups({ setMessage, loadProducts, username, password });
   const {
     backups,
-    setBackups,
-    selectedBackup,
-    setSelectedBackup,
     isBackupLoading,
-    setIsBackupLoading,
-    loadBackups,
+    createBackup,
+    downloadBackup,
+    deleteBackup,
     restoreBackup,
-    resetToDefault,
-    cleanupBackups,
-    deleteBackup
+    handleBackupFileSelect,
+    selectedBackupFile,
+    uploadAndRestoreBackup
   } = backupsApi;
 
   // Use the custom coupons hook
   const couponsApi = useAdminCoupons({ setMessage, username, password });
   const {
     coupons,
-    setCoupons,
-    isCouponLoading,
-    setIsCouponLoading,
     couponForm,
     setCouponForm,
-    loadCoupons,
+    isCouponLoading,
     createCoupon,
     deleteCoupon
   } = couponsApi;
 
+  // Local state
   const [activeTab, setActiveTab] = useState('dashboard');
   
   // Unified management states
@@ -138,16 +110,19 @@ export default function AdminPage() {
   const [imageAnalysis, setImageAnalysis] = useState(null);
   const [isImageLoading, setIsImageLoading] = useState(false);
 
+  // Auto-logout timer
   useEffect(() => {
     if (isAuthenticated) {
-      loadProducts();
-      loadBackups();
-      loadCoupons();
+      const timer = setTimeout(() => {
+        handleLogout();
+        setMessage("Session expired due to inactivity");
+      }, SESSION_TIMEOUT);
+      return () => clearTimeout(timer);
     }
   }, [isAuthenticated]);
 
   // Input sanitization
-  const sanitizeInput = (input) => {
+  const sanitizeInput = (input: string) => {
     return input.replace(/[<>]/g, '').trim();
   };
 
@@ -175,29 +150,24 @@ export default function AdminPage() {
     try {
       const response = await fetch('/api/admin/update-prices', {
         method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
+        headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({
-          username: credentials.username,
-          password: credentials.password,
-          category: selectedCategory,
           priceChangePercent: percent,
-          productId: selectedProductId !== "all" ? selectedProductId : undefined
+          credentials
         }),
       });
 
       const data = await response.json();
 
       if (response.ok) {
-        setMessage(`Successfully updated prices for ${data.updatedCount} items.`);
-        loadProducts(); // Refresh products
+        setMessage(`✅ ${data.message || 'Prices updated successfully!'}`);
+        loadProducts();
       } else {
-        setMessage(data.error || "Failed to update prices");
+        setMessage(`❌ Error: ${data.error || 'Failed to update prices'}`);
       }
     } catch (error) {
-      console.error('Error updating prices:', error);
-      setMessage("Network error. Please try again.");
+      console.error('Price update error:', error);
+      setMessage(`❌ Error: Failed to update prices. Please try again.`);
     } finally {
       setIsLoading(false);
     }
@@ -221,30 +191,27 @@ export default function AdminPage() {
     try {
       const response = await fetch('/api/admin/update-inventory', {
         method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
+        headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({
-          username: credentials.username,
-          password: credentials.password,
-          category: selectedCategory,
           stockStatus,
-          quantity: quantity ? parseInt(quantity) : null,
-          productId: selectedProductId !== "all" ? selectedProductId : undefined
+          quantity: quantity || null,
+          selectedCategory: selectedCategory === "all" ? null : selectedCategory,
+          selectedProductId: selectedProductId === "all" ? null : selectedProductId,
+          credentials
         }),
       });
 
       const data = await response.json();
 
       if (response.ok) {
-        setMessage(`Successfully updated inventory for ${data.updatedCount} items.`);
-        loadProducts(); // Refresh products
+        setMessage(`✅ ${data.message || 'Inventory updated successfully!'}`);
+        loadProducts();
       } else {
-        setMessage(data.error || "Failed to update inventory");
+        setMessage(`❌ Error: ${data.error || 'Failed to update inventory'}`);
       }
     } catch (error) {
-      console.error('Error updating inventory:', error);
-      setMessage("Network error. Please try again.");
+      console.error('Inventory update error:', error);
+      setMessage(`❌ Error: Failed to update inventory. Please try again.`);
     } finally {
       setIsLoading(false);
     }
@@ -261,29 +228,23 @@ export default function AdminPage() {
     setMessage("");
 
     try {
-      const response = await fetch('/api/admin/backup-management', {
+      const response = await fetch('/api/admin/analyze-images', {
         method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          username,
-          password,
-          action: 'analyze_images'
-        }),
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ username, password }),
       });
 
       const data = await response.json();
 
       if (response.ok) {
         setImageAnalysis(data.analysis);
-        setMessage(data.message);
+        setMessage(`✅ Analysis complete: Found ${data.analysis.unusedCount} unused images (${data.analysis.totalSize})`);
       } else {
-        setMessage(data.error || "Failed to analyze images");
+        setMessage(`❌ Error: ${data.error || 'Failed to analyze images'}`);
       }
     } catch (error) {
-      console.error('Error analyzing images:', error);
-      setMessage("Network error. Please try again.");
+      console.error('Image analysis error:', error);
+      setMessage(`❌ Error: Failed to analyze images. Please try again.`);
     } finally {
       setIsImageLoading(false);
     }
@@ -304,63 +265,38 @@ export default function AdminPage() {
     setMessage("");
 
     try {
-      const response = await fetch('/api/admin/backup-management', {
+      const response = await fetch('/api/admin/cleanup-images', {
         method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          username,
-          password,
-          action: 'cleanup_images'
-        }),
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ username, password }),
       });
 
       const data = await response.json();
 
       if (response.ok) {
-        setMessage(data.message);
-        // Refresh analysis after cleanup
-        setTimeout(() => analyzeImages(), 1000);
+        setImageAnalysis(null);
+        setMessage(`✅ Cleanup complete: ${data.message}`);
       } else {
-        setMessage(data.error || "Failed to cleanup images");
+        setMessage(`❌ Error: ${data.error || 'Failed to cleanup images'}`);
       }
     } catch (error) {
-      console.error('Error cleaning up images:', error);
-      setMessage("Network error. Please try again.");
+      console.error('Image cleanup error:', error);
+      setMessage(`❌ Error: Failed to cleanup images. Please try again.`);
     } finally {
       setIsImageLoading(false);
     }
   };
 
-  // Update categoryProducts when selectedCategory changes
+  // Update selected product when category changes
   useEffect(() => {
-    if (selectedCategory && selectedCategory !== "all" && products.length > 0) {
-      // Find all products/variants in the selected category
-      const variants = [];
-      // Find main product
-      products.forEach((p) => {
-        if (p.categorySlug === selectedCategory) {
-          variants.push({
-            id: p.id,
-            name: p.name,
-            type: "product"
-          });
-        }
-      });
-      // Find variants from productVariants
-      const allVariants = productsData.productVariants?.[selectedCategory]?.variants || [];
-      allVariants.forEach((v) => {
-        variants.push({
-          id: v.id,
-          name: v.name + (v.capacity ? ` (${v.capacity})` : ""),
-          type: "variant"
-        });
-      });
-      setCategoryProducts(variants);
-      setSelectedProductId("all");
+    if (selectedCategory && selectedCategory !== "all") {
+      const categoryProducts = products.filter(p => p.category === selectedCategory);
+      if (categoryProducts.length > 0) {
+        setSelectedProductId(categoryProducts[0].catNo);
+      } else {
+        setSelectedProductId("all");
+      }
     } else {
-      setCategoryProducts([]);
       setSelectedProductId("all");
     }
   }, [selectedCategory, products]);
@@ -372,23 +308,42 @@ export default function AdminPage() {
           <GlassCard variant="primary" padding="large" className="w-full max-w-md flex flex-col items-center text-center bg-white/20 text-white/95 shadow-2xl">
             <div className="text-center mb-8">
               <GlassIcon icon="🔐" variant="primary" size="large" />
-              <Heading as="h1" gradient="linear-gradient(to right, #009ffd, #2a2a72)" className="mt-4 mb-2" size="secondary">
-                Admin Access
-              </Heading>
-              <p className="text-white/80">Enter credentials to access admin panel</p>
+              <h1 className="text-2xl font-bold mt-4 mb-2">Admin Login</h1>
+              <p className="text-white/80">Enter your admin credentials to continue</p>
             </div>
 
-            <form onSubmit={handleLogin} className="space-y-6">
+            {/* Login Attempts Warning */}
+            {loginAttempts > 0 && loginAttempts < 3 && (
+              <div className="w-full mb-4 p-3 bg-yellow-500/20 border border-yellow-400/30 rounded-lg">
+                <p className="text-yellow-200 text-sm">
+                  ⚠️ {3 - loginAttempts} attempt{3 - loginAttempts !== 1 ? 's' : ''} remaining
+                </p>
+              </div>
+            )}
+
+            {/* Account Locked Warning */}
+            {isLocked && (
+              <div className="w-full mb-4 p-3 bg-red-500/20 border border-red-400/30 rounded-lg">
+                <p className="text-red-200 text-sm">
+                  🔒 Account locked. Try again in {Math.ceil(lockTimeRemaining / 1000)} seconds
+                </p>
+              </div>
+            )}
+
+            <form onSubmit={handleLogin} className="w-full space-y-4">
               <div>
                 <label className="block text-sm font-medium text-white/80 mb-2">Username</label>
                 <GlassInput
                   type="text"
                   value={username}
-                  onChange={(e) => setUsername(e.target.value)}
-                  placeholder="Enter username"
+                  onChange={(e) => setUsername(sanitizeInput(e.target.value))}
+                  placeholder="Username"
                   required
-                  disabled={isLocked || isLoading}
+                  disabled={isLocked}
                   maxLength={50}
+                  min={1}
+                  max={100}
+                  step={1}
                 />
               </div>
 
@@ -398,44 +353,36 @@ export default function AdminPage() {
                   type="password"
                   value={password}
                   onChange={(e) => setPassword(e.target.value)}
-                  placeholder="Enter password"
+                  placeholder="Password"
                   required
-                  disabled={isLocked || isLoading}
+                  disabled={isLocked}
                   maxLength={100}
+                  min={1}
+                  max={100}
+                  step={1}
                 />
               </div>
 
               {message && (
-                <div className={`p-3 rounded-xl text-sm ${message.includes('successful') ? 'bg-green-500/20 text-green-300' : 'bg-red-500/20 text-red-300'}`}>
+                <div className={`text-sm p-3 rounded-lg ${
+                  message.includes('✅') 
+                    ? 'bg-green-500/20 border border-green-400/30 text-green-200' 
+                    : 'bg-red-500/20 border border-red-400/30 text-red-200'
+                }`}>
                   {message}
                 </div>
               )}
 
-              <GlassButton
-                type="submit"
-                variant="accent"
-                size="large"
-                className="w-full"
-                disabled={isLocked || isLoading}
-              >
-                {isLoading ? (
-                  <span>Logging in...</span>
-                ) : isLocked ? (
-                  <span>Account Locked</span>
-                ) : (
-                  <>
-                    <span>Login</span>
-                    <span>→</span>
-                  </>
-                )}
-              </GlassButton>
-            </form>
-
-            {isLocked && (
-              <div className="mt-4 p-3 bg-red-500/20 rounded-xl text-sm text-red-300">
-                Account temporarily locked due to multiple failed attempts.
+              <div className="flex justify-center pt-4">
+                <button
+                  type="submit"
+                  disabled={isLoading || isLocked}
+                  className="glass-button glass-button--primary glass-button--large w-full"
+                >
+                  {isLoading ? 'Signing in...' : 'Sign In'}
+                </button>
               </div>
-            )}
+            </form>
           </GlassCard>
         </section>
       </AdminLayout>
@@ -444,243 +391,68 @@ export default function AdminPage() {
 
   return (
     <AdminLayout isAuthenticated={true} username={username} handleLogout={handleLogout}>
-        {/* Tab Navigation */}
-        <section className="w-full">
-          <GlassCard variant="secondary" padding="medium" className="w-full">
-            <div className="flex gap-1 md:gap-2 overflow-x-auto whitespace-nowrap scrollbar-thin scrollbar-thumb-blue-300 scrollbar-track-transparent px-1 md:px-0 justify-center md:justify-start">
-              {ADMIN_TABS.map((tab) => (
-                <GlassButton
-                  key={tab.id}
-                  onClick={() => setActiveTab(tab.id)}
-                  variant={activeTab === tab.id ? "primary" : "secondary"}
-                  size="medium"
-                  className={`flex items-center gap-2 uppercase font-bold tracking-wide px-3 py-2 md:px-5 md:py-2 rounded-2xl transition-all duration-200 min-w-[120px] ${activeTab === tab.id ? "shadow-lg shadow-blue-400/30" : ""}`}
-                >
-                  <span className="flex items-center text-lg md:text-xl">{tab.icon}</span>
-                  <span className="text-sm md:text-base">{tab.label}</span>
-                </GlassButton>
-              ))}
-            </div>
-          </GlassCard>
-        </section>
-        {/* Tab Content */}
-        <section className="w-full">
-          <GlassCard variant="secondary" padding="large" className="w-full">
-            {/* Dashboard Tab */}
-            {activeTab === 'dashboard' && (
-              <div>
-                <h2 className="text-3xl font-bold text-white mb-6 flex items-center gap-2">
-                  <span>📊</span>
-                  Dashboard
-                </h2>
-                
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6 mb-8">
-                  <GlassCard variant="primary" padding="default" className="w-full max-w-xs flex flex-col items-center text-center bg-white/20 text-white/95">
-                    <GlassIcon icon="📦" variant="primary" size="medium" className="mb-2" />
-                    <Heading as="h2" gradient="linear-gradient(to right, #009ffd, #2a2a72)" className="mb-1 text-base md:text-lg" size="secondary">
-                      Total Products
-                    </Heading>
-                    <div className="text-2xl font-bold text-white">{products.length}</div>
-                  </GlassCard>
-                  <GlassCard variant="primary" padding="default" className="w-full max-w-xs flex flex-col items-center text-center bg-white/20 text-white/95">
-                    <GlassIcon icon="🏷️" variant="primary" size="medium" className="mb-2" />
-                    <Heading as="h2" gradient="linear-gradient(to right, #009ffd, #2a2a72)" className="mb-1 text-base md:text-lg" size="secondary">
-                      Categories
-                    </Heading>
-                    <div className="text-2xl font-bold text-white">{categories.length}</div>
-                  </GlassCard>
-                  <GlassCard variant="primary" padding="default" className="w-full max-w-xs flex flex-col items-center text-center bg-white/20 text-white/95">
-                    <GlassIcon icon="💾" variant="primary" size="medium" className="mb-2" />
-                    <Heading as="h2" gradient="linear-gradient(to right, #009ffd, #2a2a72)" className="mb-1 text-base md:text-lg" size="secondary">
-                      Backups
-                    </Heading>
-                    <div className="text-2xl font-bold text-white">{backups.length}</div>
-                  </GlassCard>
-                  <GlassCard variant="primary" padding="default" className="w-full max-w-xs flex flex-col items-center text-center bg-white/20 text-white/95">
-                    <GlassIcon icon="🎫" variant="primary" size="medium" className="mb-2" />
-                    <Heading as="h2" gradient="linear-gradient(to right, #009ffd, #2a2a72)" className="mb-1 text-base md:text-lg" size="secondary">
-                      Active Coupons
-                    </Heading>
-                    <div className="text-2xl font-bold text-white">{coupons.length}</div>
-                  </GlassCard>
-                </div>
-
-                <GlassContainer>
-                  <h3 className="text-xl font-bold text-white mb-4">Quick Actions</h3>
-                  <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-                    <GlassButton
-                      onClick={() => setActiveTab('prices')}
-                      variant="accent"
-                      size="large"
-                      className="w-full"
-                    >
-                      <span>💰</span>
-                      <span>Update Prices</span>
-                    </GlassButton>
-                    <GlassButton
-                      onClick={() => setActiveTab('inventory')}
-                      variant="accent"
-                      size="large"
-                      className="w-full"
-                    >
-                      <span>📦</span>
-                      <span>Manage Inventory</span>
-                    </GlassButton>
-                    <GlassButton
-                      onClick={() => setActiveTab('add-products')}
-                      variant="accent"
-                      size="large"
-                      className="w-full"
-                    >
-                      <span>➕</span>
-                      <span>Add Products</span>
-                    </GlassButton>
-                  </div>
-                </GlassContainer>
-              </div>
-            )}
-
-            {/* Price Management Tab */}
-            {activeTab === 'prices' && (
-              <PriceManagement
-                categories={categories}
-                selectedCategory={selectedCategory}
-                setSelectedCategory={setSelectedCategory}
-                categoryProducts={categoryProducts}
-                selectedProductId={selectedProductId}
-                setSelectedProductId={setSelectedProductId}
-                priceChangePercent={priceChangePercent}
-                setPriceChangePercent={setPriceChangePercent}
-                updatePrices={updatePrices}
-                isLoading={isLoading}
-              />
-            )}
-
-            {/* Inventory Management Tab */}
-            {activeTab === 'inventory' && (
-              <InventoryManagement
-                categories={categories}
-                selectedCategory={selectedCategory}
-                setSelectedCategory={setSelectedCategory}
-                categoryProducts={categoryProducts}
-                selectedProductId={selectedProductId}
-                setSelectedProductId={setSelectedProductId}
-                stockStatus={stockStatus}
-                setStockStatus={setStockStatus}
-                quantity={quantity}
-                setQuantity={setQuantity}
-                updateInventory={updateInventory}
-                isLoading={isLoading}
-              />
-            )}
-
-            {/* Add Products Tab */}
-            {activeTab === 'add-products' && (
-              <ProductAddition
-                categoryForm={categoryForm}
-                setCategoryForm={setCategoryForm}
-                productForm={productForm}
-                setProductForm={setProductForm}
-                newDimensionField={newDimensionField}
-                setNewDimensionField={setNewDimensionField}
-                newFeature={newFeature}
-                setNewFeature={setNewFeature}
-                isAddLoading={isAddLoading}
-                isUploading={isUploading}
-                selectedImage={selectedImage}
-                imagePreview={imagePreview}
-                addCategory={addCategory}
-                addProduct={addProduct}
-                addDimensionField={addDimensionField}
-                removeDimensionField={removeDimensionField}
-                handleImageSelect={handleImageSelect}
-                handleImageUpload={handleImageUpload}
-              />
-            )}
-
-            {/* Backup Management Tab */}
-            {activeTab === 'backups' && (
-              <BackupManagement
-                backups={backups}
-                selectedBackup={selectedBackup}
-                setSelectedBackup={setSelectedBackup}
-                isBackupLoading={isBackupLoading || isImageLoading}
-                restoreBackup={restoreBackup}
-                resetToDefault={resetToDefault}
-                cleanupBackups={cleanupBackups}
-                deleteBackup={deleteBackup}
-                analyzeImages={analyzeImages}
-                cleanupImages={cleanupImages}
-                imageAnalysis={imageAnalysis}
-              />
-            )}
-
-            {/* Coupon Management Tab */}
-            {activeTab === 'coupons' && (
-              <CouponManagement
-                coupons={coupons}
-                isCouponLoading={isCouponLoading}
-                couponForm={couponForm}
-                setCouponForm={setCouponForm}
-                createCoupon={createCoupon}
-                deleteCoupon={deleteCoupon}
-              />
-            )}
-
-            {message && (
-              <div className={`p-4 rounded-xl text-sm ${message.includes('Successfully') ? 'bg-green-500/20 text-green-300' : 'bg-red-500/20 text-red-300'}`}>
-                {message}
-              </div>
-            )}
-          </GlassCard>
-        </section>
-
-        {/* Products Preview */}
-        <section className="w-full">
-          <GlassCard variant="secondary" padding="large" className="w-full">
-            <Heading as="h2" gradient="linear-gradient(to right, #009ffd, #2a2a72)" className="mb-6" size="secondary">Current Products</Heading>
-            
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-              {products.map((product) => (
-                <GlassContainer key={product.id} padding="small">
-                  <div className="flex items-center justify-between mb-2">
-                    <span className="text-xs font-medium text-white/60 uppercase tracking-wider bg-white/10 px-2 py-1 rounded-full">
-                      {product.category}
-                    </span>
-                    <span className="text-xs text-white/60">{product.catNo}</span>
-                  </div>
-                  <h3 className="text-lg font-bold text-white mb-2">{product.name}</h3>
-                  <p className="text-sm text-white/70 mb-3">{product.description}</p>
-                  
-                  {/* Stock Status */}
-                  <div className="mb-3">
-                    <StockStatusBadge status={product.stockStatus || 'in_stock'} />
-                    {product.quantity !== undefined && product.quantity !== null && (
-                      <span className="ml-2 text-sm text-white/60">
-                        Qty: {product.quantity}
-                      </span>
-                    )}
-                  </div>
-                  
-                  <div className="space-y-1 text-sm">
-                    <div className="flex justify-between">
-                      <span className="text-white/80">Price Range:</span>
-                      <span className="text-white font-medium">{product.priceRange}</span>
-                    </div>
-                    <div className="flex justify-between">
-                      <span className="text-white/80">Capacity:</span>
-                      <span className="text-white">{product.capacity}</span>
-                    </div>
-                    <div className="flex justify-between">
-                      <span className="text-white/80">Packaging:</span>
-                      <span className="text-white">{product.packaging}</span>
-                    </div>
-                  </div>
-                </GlassContainer>
-              ))}
-            </div>
-          </GlassCard>
-        </section>
+      <AdminTabNavigation 
+        tabs={ADMIN_TABS}
+        activeTab={activeTab}
+        onTabChange={setActiveTab}
+      />
+      
+      <AdminTabContent
+        activeTab={activeTab}
+        onTabChange={setActiveTab}
+        products={products}
+        categories={categories}
+        backups={backups}
+        coupons={coupons}
+        isAuthenticated={isAuthenticated}
+        username={username}
+        password={password}
+        setMessage={setMessage}
+        priceChangePercent={priceChangePercent}
+        setPriceChangePercent={setPriceChangePercent}
+        stockStatus={stockStatus}
+        setStockStatus={setStockStatus}
+        quantity={quantity}
+        setQuantity={setQuantity}
+        selectedCategory={selectedCategory}
+        setSelectedCategory={setSelectedCategory}
+        selectedProductId={selectedProductId}
+        setSelectedProductId={setSelectedProductId}
+        isLoading={isLoading}
+        setIsLoading={setIsLoading}
+        updatePrices={updatePrices}
+        updateInventory={updateInventory}
+        loadProducts={loadProducts}
+        imageAnalysis={imageAnalysis}
+        setImageAnalysis={setImageAnalysis}
+        isImageLoading={isImageLoading}
+        setIsImageLoading={setIsImageLoading}
+        analyzeImages={analyzeImages}
+        cleanupImages={cleanupImages}
+        categoryForm={categoryForm}
+        setCategoryForm={setCategoryForm}
+        productForm={productForm}
+        setProductForm={setProductForm}
+        newDimensionField={newDimensionField}
+        setNewDimensionField={setNewDimensionField}
+        newFeature={newFeature}
+        setNewFeature={setNewFeature}
+        isAddLoading={isAddLoading}
+        isUploading={isUploading}
+        selectedImage={selectedImage}
+        imagePreview={imagePreview}
+        addCategory={addCategory}
+        addProduct={addProduct}
+        addDimensionField={addDimensionField}
+        removeDimensionField={removeDimensionField}
+        handleImageSelect={handleImageSelect}
+        handleImageUpload={handleImageUpload}
+        couponForm={couponForm}
+        setCouponForm={setCouponForm}
+        isCouponLoading={isCouponLoading}
+        createCoupon={createCoupon}
+        deleteCoupon={deleteCoupon}
+      />
     </AdminLayout>
   );
-} 
+}
